feat: finished the owner UI portion, removed images from book, and some other stuff that I don't remember

+++ sql/views/sale_per_day.sql
@@ -1,5 +1,5 @@
 create or replace function sale_on_day(day date)
-	returns decimal(10,2)
+	returns float
 	language plpgsql
 as
 $$
